(24/05/10 - 1) 클로즈 베타 테스트 버전 배포합니다.

+++ src/main/java/com/coverflow/question/application/QuestionServiceImpl.java
@@ -1,6 +1,5 @@
 package com.coverflow.question.application;
 
-import com.coverflow.company.domain.Company;
 import com.coverflow.company.infrastructure.CompanyRepository;
 import com.coverflow.member.application.MemberServiceImpl;
 import com.coverflow.question.domain.Question;
@@ -24,10 +23,10 @@ import java.util.ArrayList;
 import java.util.Optional;
 import java.util.UUID;
 
-import static com.coverflow.company.exception.CompanyException.CompanyNotFoundException;
 import static com.coverflow.global.constant.Constant.LARGE_PAGE_SIZE;
 import static com.coverflow.global.constant.Constant.NORMAL_PAGE_SIZE;
 import static com.coverflow.global.util.PageUtil.generatePageDesc;
+import static com.coverflow.question.exception.AnswerException.AnswerExistException;
 import static com.coverflow.question.exception.QuestionException.AlreadySelectedQuestionException;
 import static com.coverflow.question.exception.QuestionException.QuestionNotFoundException;
 
@@ -120,12 +119,8 @@ public class QuestionServiceImpl implements QuestionService {
             final SaveQuestionRequest request,
             final String memberId
     ) {
-        Company company = companyRepository.findById(request.companyId())
-                .orElseThrow(() -> new CompanyNotFoundException(request.companyId()));
-
         memberService.writeQuestion(memberId, request.reward());
         questionRepository.save(new Question(request, memberId));
-        company.updateQuestionCount(company.getQuestionCount() + 1);
     }
 
     @Override
@@ -136,6 +131,10 @@ public class QuestionServiceImpl implements QuestionService {
     ) {
         Question question = questionRepository.findById(questionId)
                 .orElseThrow(() -> new QuestionNotFoundException(questionId));
+
+        if (!question.getAnswers().isEmpty()) {
+            throw new AnswerExistException(questionId);
+        }
 
         if (question.isSelectionStatus()) {
             throw new AlreadySelectedQuestionException(questionId);
@@ -149,6 +148,10 @@ public class QuestionServiceImpl implements QuestionService {
     public void delete(final long questionId) {
         Question question = questionRepository.findById(questionId)
                 .orElseThrow(() -> new QuestionNotFoundException(questionId));
+
+        if (!question.getAnswers().isEmpty()) {
+            throw new AnswerExistException(questionId);
+        }
 
         questionRepository.delete(question);
     }
